backend/adser/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
aiokafka==0.12.0
boto3==1.35.0
httpx==0.27.2
pyjwt==2.8.0
pytest==8.3.3
pytest-asyncio==1.3.0
